E2vid integration refactor (#136)
* delete timers.py, examples/reconstruction_e2vid.py
* Restructure folders: e2vid
* Refactor module structure
* Fix file name for the test and examples
* Delete 
* Fix tests, mypy, typeguard
