+++ src/evlib/processing/__init__.py
@@ -1,1 +1,1 @@
-"""Event Vision Library."""
+"""Processing module."""
