+++ src/evlib/processing/reconstruction/__init__.py
@@ -1,1 +1,3 @@
+"""Algorithms for (intensity) image reconstruction."""
+
 from .e2vid import E2Vid as E2Vid
