--- src/evlib/processing/reconstruction/e2vid/__init__.py
@@ -1,1 +1,0 @@
-from .e2vid import E2Vid as E2Vid
