+++ src/evlib/processing/reconstruction/e2vid_module/model/model.py
@@ -2,7 +2,7 @@ from typing import Optional, List, Tuple
 
 import torch
 
-from ..base import BaseModel
+from ..base_model import BaseModel
 from ..model.unet import UNetRecurrent
 
 
@@ -36,7 +36,7 @@ class BaseE2VID(BaseModel):
         try:
             self.norm = str(config['norm'])
         except KeyError:
-            self.norm = None
+            self.norm = None  # type: ignore
 
         try:
             self.use_upsample_conv = bool(config['use_upsample_conv'])
@@ -68,7 +68,7 @@ class E2VIDRecurrent(BaseE2VID):
                                            norm=self.norm,
                                            use_upsample_conv=self.use_upsample_conv)
 
-    def forward(self, event_tensor: torch.Tensor,
+    def forward(self, event_tensor: torch.Tensor,  # type: ignore
                 prev_states: Optional[List[torch.Tensor]]) -> Tuple[torch.Tensor,
                                                                     List[torch.Tensor]]:
         """
