+++ src/evlib/processing/reconstruction/e2vid_module/utils/inference_utils.py
@@ -9,8 +9,6 @@ import cv2
 import scipy.stats as st
 import torch.nn.functional as F
 
-from .timers import Timer, CudaTimer
-
 
 def gkern(kernlen: int=5, nsig: float=1.0) -> torch.Tensor:
     """Returns a 2D Gaussian kernel array."""
@@ -39,24 +37,11 @@ class EventPreprocessor:
         else:
             print('Will normalize event tensors.')
 
-        self.hot_pixel_locations = []
-        if options.hot_pixels_file:
-            try:
-                self.hot_pixel_locations = np.loadtxt(options.hot_pixels_file, delimiter=',').astype(np.int)
-                print('Will remove {} hot pixels'.format(self.hot_pixel_locations.shape[0]))
-            except IOError:
-                print('WARNING: could not load hot pixels file: {}'.format(options.hot_pixels_file))
-
         self.flip = options.flip
         if self.flip:
             print('Will flip event tensors.')
 
     def __call__(self, events: torch.Tensor) -> torch.Tensor:
-
-        # Remove (i.e. zero out) the hot pixels
-        for x, y in self.hot_pixel_locations:
-            events[:, :, y, x] = 0
-
         # Flip tensor vertically and horizontally
         if self.flip:
             events = torch.flip(events, dims=[2, 3])
@@ -64,17 +49,16 @@ class EventPreprocessor:
         # Normalize the event tensor (voxel grid) so that
         # the mean and stddev of the nonzero values in the tensor are equal to (0.0, 1.0)
         if not self.no_normalize:
-            with CudaTimer('Normalization'):
-                nonzero_ev = (events != 0)
-                num_nonzeros = nonzero_ev.sum()
-                if num_nonzeros > 0:
-                    # compute mean and stddev of the **nonzero** elements of the event tensor
-                    # we do not use PyTorch's default mean() and std() functions since it's faster
-                    # to compute it by hand than applying those funcs to a masked array
-                    mean = events.sum() / num_nonzeros
-                    stddev = torch.sqrt((events ** 2).sum() / num_nonzeros - mean ** 2)
-                    mask = nonzero_ev.float()
-                    events = mask * (events - mean) / stddev
+            nonzero_ev = (events != 0)
+            num_nonzeros = nonzero_ev.sum()
+            if num_nonzeros > 0:
+                # compute mean and stddev of the **nonzero** elements of the event tensor
+                # we do not use PyTorch's default mean() and std() functions since it's faster
+                # to compute it by hand than applying those funcs to a masked array
+                mean = events.sum() / num_nonzeros
+                stddev = torch.sqrt((events ** 2).sum() / num_nonzeros - mean ** 2)
+                mask = nonzero_ev.float()
+                events = mask * (events - mean) / stddev
 
         return events
 
@@ -88,7 +72,7 @@ class IntensityRescaler:
 
     def __init__(self, options: Any) -> None:
         self.auto_hdr = options.auto_hdr
-        self.intensity_bounds = deque()
+        self.intensity_bounds = deque()  # type: ignore
         self.auto_hdr_median_filter_size = options.auto_hdr_median_filter_size
         self.Imin = options.Imin
         self.Imax = options.Imax
@@ -98,26 +82,24 @@ class IntensityRescaler:
         param img: [1 x 1 x H x W] Tensor taking values in [0, 1]
         """
         if self.auto_hdr:
-            with CudaTimer('Compute Imin/Imax (auto HDR)'):
-                Imin = torch.min(img).item()
-                Imax = torch.max(img).item()
+            Imin = torch.min(img).item()
+            Imax = torch.max(img).item()
 
-                # ensure that the range is at least 0.1
-                Imin = np.clip(Imin, 0.0, 0.45)
-                Imax = np.clip(Imax, 0.55, 1.0)
+            # ensure that the range is at least 0.1
+            Imin = np.clip(Imin, 0.0, 0.45)
+            Imax = np.clip(Imax, 0.55, 1.0)
 
-                # adjust image dynamic range (i.e. its contrast)
-                if len(self.intensity_bounds) > self.auto_hdr_median_filter_size:
-                    self.intensity_bounds.popleft()
+            # adjust image dynamic range (i.e. its contrast)
+            if len(self.intensity_bounds) > self.auto_hdr_median_filter_size:
+                self.intensity_bounds.popleft()
 
-                self.intensity_bounds.append((Imin, Imax))
-                self.Imin = np.median([rmin for rmin, rmax in self.intensity_bounds])
-                self.Imax = np.median([rmax for rmin, rmax in self.intensity_bounds])
+            self.intensity_bounds.append((Imin, Imax))
+            self.Imin = np.median([rmin for rmin, rmax in self.intensity_bounds])
+            self.Imax = np.median([rmax for rmin, rmax in self.intensity_bounds])
 
-        with CudaTimer('Intensity rescaling'):
-            img = 255.0 * (img - self.Imin) / (self.Imax - self.Imin)
-            img.clamp_(0.0, 255.0)
-            img = img.byte()  # convert to 8-bit tensor
+        img = 255.0 * (img - self.Imin) / (self.Imax - self.Imin)
+        img.clamp_(0.0, 255.0)
+        img = img.byte()  # convert to 8-bit tensor
 
         return img
 
@@ -135,10 +117,9 @@ class UnsharpMaskFilter:
 
     def __call__(self, img: torch.Tensor) -> torch.Tensor:
         if self.unsharp_mask_amount > 0:
-            with CudaTimer('Unsharp mask'):
-                blurred = F.conv2d(img, self.gaussian_kernel,
-                                   padding=self.gaussian_kernel_size // 2)
-                img = (1 + self.unsharp_mask_amount) * img - self.unsharp_mask_amount * blurred
+            blurred = F.conv2d(img, self.gaussian_kernel,
+                                padding=self.gaussian_kernel_size // 2)
+            img = (1 + self.unsharp_mask_amount) * img - self.unsharp_mask_amount * blurred
         return img
 
 
@@ -152,11 +133,11 @@ class ImageFilter:
 
     def __call__(self, img: torch.Tensor) -> torch.Tensor:
         if self.bilateral_filter_sigma:
-            with Timer('Bilateral filter (sigma={:.2f})'.format(self.bilateral_filter_sigma)):
-                filtered_img = np.zeros_like(img)
-                filtered_img = cv2.bilateralFilter(
-                    img, 5, 25.0 * self.bilateral_filter_sigma, 25.0 * self.bilateral_filter_sigma)
-                img = filtered_img
+            # TODO check - this looks like returning numpy, not torch.
+            filtered_img = np.zeros_like(img)
+            filtered_img = cv2.bilateralFilter(
+                img, 5, 25.0 * self.bilateral_filter_sigma, 25.0 * self.bilateral_filter_sigma)
+            img = filtered_img  # type: ignore
         return img
